Remove create/destroy and config typedef
I have had a change in philosophy since I wrote
some of this. A struct that is instanced and meant
to be accessed like a struct should not be a typedef.
I also decided to remove all malloc calls and the
context object can now be instanced directly and allocated
staically, as is more appropiate for embedded code.